+++ gobrs-async-test/src/main/java/com/gobrs/async/test/transaction/TransactionTaskA.java
@@ -1,4 +1,4 @@
-package com.gobrs.async.test.task.transaction;
+package com.gobrs.async.test.transaction;
 
 import com.gobrs.async.core.TaskSupport;
 import com.gobrs.async.core.anno.Task;
