+++ gobrs-async-test/src/main/java/com/gobrs/async/test/transaction/TransactionTaskB.java
@@ -1,4 +1,4 @@
-package com.gobrs.async.test.task.transaction;
+package com.gobrs.async.test.transaction;
 
 import com.gobrs.async.core.TaskSupport;
 import com.gobrs.async.core.anno.Task;
@@ -6,23 +6,23 @@ import com.gobrs.async.core.task.AsyncTask;
 
 /**
  * @program: gobrs-async
- * @ClassName CaseOneTaskC
+ * @ClassName GobrsTaskB
  * @description:
  * @author: sizegang
  * @create: 2022-10-31
  **/
 @Task
-public class TransactionTaskC extends AsyncTask {
+public class TransactionTaskB extends AsyncTask {
 
     @Override
     public Object task(Object o, TaskSupport support) {
-        System.out.println(" TransactionTaskC 任务执行");
-        return "CResult";
+        System.out.println(" TransactionTaskB  任务执行");
+        return "BResult";
     }
 
 
     @Override
     public void rollback(Object o) {
-        System.out.println("TransactionTaskC 开始执行回滚");
+        System.out.println("TransactionTaskB 开始执行回滚");
     }
 }
