+++ gobrs-async-test/src/test/java/com/gobrs/async/test/transaction/CaseTransaction.java
@@ -1,4 +1,4 @@
-package com.gobrs.async.test.task.transaction;
+package com.gobrs.async.test.transaction;
 
 import com.gobrs.async.core.GobrsAsync;
 import com.gobrs.async.core.common.domain.AsyncResult;
